made a clientdialog.js for the dialog modal
